Create ProjectIdea.txt
The deadline is Thursday night, September 26.

Increase by
- The general purpose of the software.
- The intended audience of the software.
- The set of features offered by the software.
- A very brief description of the type of data that would be managed by your software, if applicable.
Need to add content
Any other information you think is relevant to your project idea.
You may also submit a poster if you have prepared one already. The poster is not required at this point,  but will be later in the term.